requirements.txt
annotated-types==0.7.0
async-timeout==4.0.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.8
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
zipp==3.20.1
